requirements.txt
Pillow>=3.3.1